+++ physical_ai_server/physical_ai_server/communication/communicator.py
@@ -14,14 +14,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# Author: Dongyun Kim
+# Author: Dongyun Kim, Seongwoo Kim
 
 from functools import partial
 from typing import Any, Dict, Optional, Set, Tuple
 
 from geometry_msgs.msg import Twist
 from nav_msgs.msg import Odometry
-from physical_ai_interfaces.msg import TaskStatus
+from physical_ai_interfaces.msg import TaskStatus, TrainingStatus
 from physical_ai_interfaces.srv import (
     GetImageTopicList
 )
@@ -178,6 +178,12 @@ class Communicator:
             self.PUB_QOS_SIZE
         )
 
+        self.training_status_publisher = self.node.create_publisher(
+            TrainingStatus,
+            '/training/status',
+            self.PUB_QOS_SIZE
+        )
+
         self.heartbeat_publisher = self.node.create_publisher(
             Empty,
             'heartbeat',
@@ -269,3 +275,6 @@ class Communicator:
     def heartbeat_timer_callback(self):
         heartbeat_msg = Empty()
         self.heartbeat_publisher.publish(heartbeat_msg)
+
+    def publish_training_status(self, status: TrainingStatus):
+        self.training_status_publisher.publish(status)