+++ physical_ai_manager/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package physical_ai_manager
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.6.0 (2025-07-23)
+------------------
+* Added a new training page for training imitation learning models
+* Contributors: Kiwoong Park
+
 0.5.13 (2025-07-21)
 ------------------
 * None