+++ physical_ai_tools/CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog for package physical_ai_tools
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.6.0 (2025-07-23)
+------------------
+* Added training-related service types and message types
+* Added a new training page for training imitation learning models
+* Implemented a Training Manager to support model training through the Web UI
+* Contributors: Kiwoong Park, Seongwoo Kim
+
 0.5.13 (2025-07-21)
 ------------------
 * Updated Docker volume mount paths from `/root` to `${HOME}` for improved portability and compatibility